+++ src/components/dialog/dialog.ts
@@ -1,5 +1,5 @@
 import { LitElement, html } from 'lit';
-import { customElement, property, query, state } from 'lit/decorators.js';
+import { customElement, property, query } from 'lit/decorators.js';
 import { classMap } from 'lit/directives/class-map.js';
 import { ifDefined } from 'lit/directives/if-defined.js';
 import { animateTo, stopAnimations } from '../../internal/animate';
@@ -7,7 +7,7 @@ import { emit } from '../../internal/event';
 import { watch } from '../../internal/watch';
 import { waitForEvent } from '../../internal/event';
 import { lockBodyScrolling, unlockBodyScrolling } from '../../internal/scroll';
-import { hasSlot } from '../../internal/slot';
+import { HasSlotController } from '../../internal/slot';
 import { isPreventScrollSupported } from '../../internal/support';
 import Modal from '../../internal/modal';
 import { setDefaultAnimation, getAnimation } from '../../utilities/animation-registry';
@@ -65,10 +65,9 @@ export default class SlDialog extends LitElement {
   @query('.dialog__panel') panel: HTMLElement;
   @query('.dialog__overlay') overlay: HTMLElement;
 
+  private hasSlotController = new HasSlotController(this, ['footer']);
   private modal: Modal;
   private originalTrigger: HTMLElement | null;
-
-  @state() private hasFooter = false;
 
   /** Indicates whether or not the dialog is open. You can use this in lieu of the show/hide methods. */
   @property({ type: Boolean, reflect: true }) open = false;
@@ -87,9 +86,7 @@ export default class SlDialog extends LitElement {
 
   connectedCallback() {
     super.connectedCallback();
-
     this.modal = new Modal(this);
-    this.handleSlotChange();
   }
 
   firstUpdated() {
@@ -208,10 +205,6 @@ export default class SlDialog extends LitElement {
     }
   }
 
-  handleSlotChange() {
-    this.hasFooter = hasSlot(this, 'footer');
-  }
-
   render() {
     return html`
       <div
@@ -219,7 +212,7 @@ export default class SlDialog extends LitElement {
         class=${classMap({
           dialog: true,
           'dialog--open': this.open,
-          'dialog--has-footer': this.hasFooter
+          'dialog--has-footer': this.hasSlotController.test('footer')
         })}
         @keydown=${this.handleKeyDown}
       >
@@ -257,7 +250,7 @@ export default class SlDialog extends LitElement {
           </div>
 
           <footer part="footer" class="dialog__footer">
-            <slot name="footer" @slotchange=${this.handleSlotChange}></slot>
+            <slot name="footer"></slot>
           </footer>
         </div>
       </div>